Eagerly mark vm->mark_object_array
Objects are added to `vm->mark_object_array` via
`rb_gc_register_mark_object`.  Objects passed to
`rb_gc_register_mark_object` are expected not to move.  C extensions
typically create the object passed to that function, then hold a
reference to the object.  Since the C extensions are holding a
reference, we can't allow the address of the object to change.

This commit eagerly iterates over the mark_object_array so that objects
inside that array are considered roots, and that will keep them pinned.