Don't verify roots, ignore T_MOVED in mark maybe
Objects on the C stack can get moved.  They get on the C stack when we
are actually executing the movement code.  That means the pointer could
be moved and left on the C stack, but is not visible from Ruby land.  If
those pointers get moved, we should not worry.