Use shared substrings in feature index cache hash
Before this patch, `features_index_add` would use `rb_str_subseq` to get
a substring of the feature being added to the loaded features list.
`features_index_add_single` would use `ruby_strdup` to copy that string
and use it as a hash key in `loaded_features_index`.

This patch changes `features_index_add` to index in to the underlying
character array stored in the Ruby string, and use that as the hash key
without copying its contents.  The cache also needs keys that do not
contain file extensions, so this patch will allocate one new string that
does not contain the file extension, then indexes in to that character
array rather than use substrings.

The strings that do not have the file extension are added to a new array
on the VM `loaded_features_index_pool` to ensure liveness.  The loaded
features array already ensures liveness of the strings *with* file
extensions.